+++ backend/src/main/java/ca/metricalsky/winston/service/fetch/action/FetchRepliesActionHandler.java
@@ -3,8 +3,6 @@ package ca.metricalsky.winston.service.fetch.action;
 import ca.metricalsky.winston.client.YouTubeClientAdapter;
 import ca.metricalsky.winston.dto.CommentDto;
 import ca.metricalsky.winston.entity.fetch.FetchAction;
-import ca.metricalsky.winston.events.FetchEvent;
-import ca.metricalsky.winston.events.FetchStatus;
 import ca.metricalsky.winston.mapper.dto.CommentDtoMapper;
 import ca.metricalsky.winston.mapper.entity.CommentMapper;
 import ca.metricalsky.winston.service.CommentService;
@@ -58,11 +56,5 @@ public class FetchRepliesActionHandler extends FetchActionHandler<CommentDto> {
                 .objectId(fetchAction.getObjectId())
                 .pageToken(youTubeResponse.getNextPageToken())
                 .build();
-    }
-
-    @Override
-    protected FetchEvent getFetchEvent(FetchResult<CommentDto> fetchResult) {
-        var status = fetchResult.hasNextFetchAction() ? FetchStatus.FETCHING : FetchStatus.COMPLETED;
-        return FetchEvent.data("fetch-replies", fetchResult.objectId(), status, fetchResult.items());
     }
 }