+++ backend/src/main/java/ca/metricalsky/winston/service/fetch/request/FetchVideoRepliesRequestHandler.java
@@ -3,6 +3,7 @@ package ca.metricalsky.winston.service.fetch.request;
 import ca.metricalsky.winston.entity.fetch.FetchAction;
 import ca.metricalsky.winston.entity.fetch.FetchRequest;
 import ca.metricalsky.winston.events.SsePublisher;
+import ca.metricalsky.winston.repository.CommentRepository;
 import ca.metricalsky.winston.service.fetch.FetchRequestService;
 import ca.metricalsky.winston.service.fetch.action.FetchActionHandlerFactory;
 import lombok.RequiredArgsConstructor;
@@ -10,19 +11,19 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class DefaultFetchRequestHandler implements FetchRequestHandler {
+public class FetchVideoRepliesRequestHandler implements FetchRequestHandler {
 
+    private final CommentRepository commentRepository;
     private final FetchActionHandlerFactory fetchActionHandlerFactory;
     private final FetchRequestService fetchRequestService;
 
     @Override
     public void fetch(FetchRequest fetchRequest, SsePublisher ssePublisher) {
         fetchRequest = fetchRequestService.startFetch(fetchRequest);
-        var fetchAction = getFirstFetchAction(fetchRequest);
         try {
-            while (fetchAction != null) {
-                var actionHandler = fetchActionHandlerFactory.getHandlerForAction(fetchAction);
-                fetchAction = actionHandler.fetch(fetchAction, ssePublisher);
+            var videoId = fetchRequest.getObjectId();
+            for (var commentId : commentRepository.findIdsMissingRepliesByVideoId(videoId)) {
+                fetchReplies(fetchRequest, commentId, ssePublisher);
             }
             fetchRequestService.fetchCompleted(fetchRequest);
         } catch (RuntimeException ex) {
@@ -31,13 +32,19 @@ public class DefaultFetchRequestHandler implements FetchRequestHandler {
         }
     }
 
-    private static FetchAction getFirstFetchAction(FetchRequest fetchRequest) {
+    private void fetchReplies(FetchRequest fetchRequest, String commentId, SsePublisher ssePublisher) {
+        var fetchAction = getFirstFetchAction(fetchRequest, commentId);
+        while (fetchAction != null) {
+            var actionHandler = fetchActionHandlerFactory.getHandlerForAction(fetchAction);
+            fetchAction = actionHandler.fetch(fetchAction, ssePublisher);
+        }
+    }
+
+    private static FetchAction getFirstFetchAction(FetchRequest fetchRequest, String commentId) {
         return FetchAction.builder()
                 .fetchRequestId(fetchRequest.getId())
-                .actionType(FetchAction.ActionType.valueOf(fetchRequest.getFetchType().name()))
-                .objectId(fetchRequest.getObjectId())
-                .publishedAfter(fetchRequest.getPublishedAfter())
-                .publishedBefore(fetchRequest.getPublishedBefore())
+                .actionType(FetchAction.ActionType.REPLIES)
+                .objectId(commentId)
                 .build();
     }
 }