+++ frontend/src/routes/channels/FetchChannelAction.tsx
@@ -2,10 +2,10 @@ import {useFetchChannelByHandleMutation} from "../../store/slices/api";
 import {NotificationsSource} from "../../components/NotificationsSource";
 import {useNavigate} from "react-router";
 import {EventSourceProvider} from "react-sse-hooks";
-import {FetchChannelEvent} from "../../model/events/FetchEvent";
+import {FetchChannelEvent, FetchStatusEvent} from "../../model/events/FetchEvent";
 import {useAppDispatch} from "../../store/hooks";
 import {appendFetchedChannels} from "../../store/slices/channels";
-
+import {updateFetchStatus} from "../../store/slices/fetches";
 type FetchChannelActionProps = {
     channelHandle: string
 }
@@ -21,19 +21,27 @@ export const FetchChannelAction = ({channelHandle}: FetchChannelActionProps) => 
         fetchChannelsByHandle({ subscriptionId, channelHandle });
     }
 
-    const handleEvent = (event: FetchChannelEvent) => {
-        if (event.status === 'COMPLETED') {
-            dispatch(appendFetchedChannels(event.items));
+    const handleDataEvent = (event: FetchChannelEvent) => {
+        dispatch(appendFetchedChannels(event.items));
+        if (event.items.length > 0) {
             navigate(`/channels/${event.items[0].id}`);
         }
+    }
+
+    const handleStatusEvent = (event: FetchStatusEvent) => {
+        dispatch(updateFetchStatus({
+            fetchType: "channel",
+            objectId: channelHandle,
+            status: event.status
+        }));
     }
 
     return (
         <EventSourceProvider>
             <NotificationsSource
                 onSubscribed={handleSubscribed}
-                eventName={"fetch-channels"}
-                onEvent={handleEvent}
+                onDataEvent={handleDataEvent}
+                onStatusEvent={handleStatusEvent}
             />
         </EventSourceProvider>
     );