+++ frontend/src/routes/videos/id/FetchCommentsAction.tsx
@@ -1,10 +1,11 @@
 import {useFetchCommentsByVideoIdMutation} from "../../../store/slices/api";
-import {commentsAdapter, commentsApiUtils} from "../../../store/slices/comments";
+import {commentsAdapter, commentsApiUtils, repliesAdapter} from "../../../store/slices/comments";
 import {useAppDispatch} from "../../../store/hooks";
 import {EventSourceProvider} from "react-sse-hooks";
 import {NotificationsSource} from "../../../components/NotificationsSource";
-import {fetchedComments} from "../../../store/slices/fetches";
-import {FetchCommentsEvent} from "../../../model/events/FetchEvent";
+import {fetchedComments, updateFetchStatus} from "../../../store/slices/fetches";
+import {FetchCommentsEvent, FetchStatusEvent} from "../../../model/events/FetchEvent";
+import {videosApiUtils} from "../../../store/slices/videos";
 
 type FetchVideosActionProps = {
     videoId: string,
@@ -19,12 +20,29 @@ export const FetchCommentsAction = ({videoId}: FetchVideosActionProps) => {
         fetchCommentsByVideoId({subscriptionId, videoId});
     }
 
-    const handleEvent = (event: FetchCommentsEvent) => {
-        dispatch(fetchedComments(event));
-        if (event.status !== 'FAILED') {
-            dispatch(commentsApiUtils.updateQueryData("listCommentsByVideoId", videoId, draft => {
-                commentsAdapter.addMany(draft, event.items);
+    const handleDataEvent = (event: FetchCommentsEvent) => {
+        dispatch(commentsApiUtils.updateQueryData("listCommentsByVideoId", videoId, draft => {
+            const comments = event.items.map(comment => ({
+                ...comment,
+                replies: repliesAdapter.addMany(repliesAdapter.getInitialState(), comment.replies)
             }))
+            commentsAdapter.addMany(draft, comments);
+        }))
+        dispatch(fetchedComments(event));
+    }
+
+    const handleStatusEvent = (event: FetchStatusEvent) => {
+        dispatch(updateFetchStatus({
+            fetchType: "comments",
+            objectId: videoId,
+            status: event.status,
+        }))
+        if (event.status === "FAILED") {
+            if (event.error.type === "/api/problem/comments-disabled") {
+                dispatch(videosApiUtils.updateQueryData("findVideoById", videoId, draft => {
+                    draft.commentsDisabled = true
+                }))
+            }
         }
     }
 
@@ -32,8 +50,8 @@ export const FetchCommentsAction = ({videoId}: FetchVideosActionProps) => {
         <EventSourceProvider>
             <NotificationsSource
                 onSubscribed={handleSubscribed}
-                eventName={"fetch-comments"}
-                onEvent={handleEvent}
+                onDataEvent={handleDataEvent}
+                onStatusEvent={handleStatusEvent}
             />
         </EventSourceProvider>
     )