+++ frontend/src/routes/videos/id/FetchVideoRepliesAction.tsx
@@ -1,28 +1,27 @@
 import { useAppDispatch } from "../../../store/hooks";
-import { useFetchRepliesByCommentIdMutation} from "../../../store/slices/api";
+import {useFetchRepliesByVideoIdMutation} from "../../../store/slices/api";
 import {FetchCommentsEvent, FetchStatusEvent} from "../../../model/events/FetchEvent";
 import {EventSourceProvider} from "react-sse-hooks";
 import {NotificationsSource} from "../../../components/NotificationsSource";
-import {fetchedReplies, updateFetchStatus} from "../../../store/slices/fetches";
+import {updateFetchStatus} from "../../../store/slices/fetches";
 import {commentsAdapter, commentsApiUtils, repliesAdapter} from "../../../store/slices/comments";
 
 type FetchRepliesActionProps = {
-    commentId: string;
+    videoId: string;
 }
 
-export const FetchRepliesAction = ({commentId}: FetchRepliesActionProps) => {
+export const FetchVideoRepliesAction = ({videoId}: FetchRepliesActionProps) => {
 
-    const [fetchRepliesByCommentId] = useFetchRepliesByCommentIdMutation();
+    const [fetchRepliesByVideoId] = useFetchRepliesByVideoIdMutation();
     const dispatch = useAppDispatch();
 
     const handleSubscribed = (subscriptionId: string) => {
-        fetchRepliesByCommentId({subscriptionId, commentId});
+        fetchRepliesByVideoId({subscriptionId, videoId});
     }
 
     const handleDataEvent = (event: FetchCommentsEvent) => {
-        dispatch(fetchedReplies(event));
         if (event.items.length > 0) {
-            const videoId = event.items[0].videoId;
+            const commentId = event.objectId;
             dispatch(commentsApiUtils.updateQueryData("listCommentsByVideoId", videoId, draft => {
                 const comment = commentsAdapter.getSelectors().selectById(draft, commentId)
                 commentsAdapter.setOne(draft, {
@@ -36,7 +35,7 @@ export const FetchRepliesAction = ({commentId}: FetchRepliesActionProps) => {
     const handleStatusEvent = (event: FetchStatusEvent) => {
         dispatch(updateFetchStatus({
             fetchType: "replies",
-            objectId: commentId,
+            objectId: videoId,
             status: event.status,
         }))
     }