+++ frontend/src/store/slices/fetches.ts
@@ -1,30 +1,26 @@
 import {createSlice, PayloadAction} from "@reduxjs/toolkit";
-import {VideoWithChannelIdDto} from "../../model/VideoDto";
 import {ChannelDto} from "../../model/ChannelDto";
-import {CommentDto} from "../../model/CommentDto";
 import {FetchCommentsEvent, FetchVideosEvent} from "../../model/events/FetchEvent";
-import {ProblemDetail} from "../../model/events/ProblemDetail";
 
-export type FetchState<T> = {
+export type FetchState = {
     id: string;
     mode?: 'ALL' | 'LATEST';
     status: 'READY' | 'REQUESTED' | 'FETCHING' | 'COMPLETED' | 'FAILED';
-    data?: T[];
-    error?: ProblemDetail;
+    count: number
 }
 
 type FetchStates = {
     channel: {
-        [id: string]: FetchState<ChannelDto>;
+        [id: string]: FetchState;
     },
     videos: {
-        [id: string]: FetchState<VideoWithChannelIdDto>;
+        [id: string]: FetchState;
     },
     comments: {
-        [id: string]: FetchState<CommentDto>;
+        [id: string]: FetchState;
     },
     replies: {
-        [id: string]: FetchState<CommentDto>;
+        [id: string]: FetchState;
     }
 }
 
@@ -40,6 +36,12 @@ type FetchVideosRequest = {
     mode: 'ALL' | 'LATEST';
 }
 
+type UpdateFetchStatus = {
+    fetchType: keyof FetchStates;
+    objectId: string;
+    status: 'COMPLETED' | 'FAILED';
+}
+
 export const fetchesSlice = createSlice({
     name: "fetches",
     initialState,
@@ -48,12 +50,14 @@ export const fetchesSlice = createSlice({
             state.channel[action.payload] = {
                 id: action.payload,
                 status: 'REQUESTED',
+                count: 0
             }
         },
         initFetchStateForChannel: (state, action: PayloadAction<ChannelDto>) => {
             state.videos[action.payload.id] = {
                 id: action.payload.id,
                 status: 'READY',
+                count: 0
             }
         },
         requestedVideosForChannelId: (state, action: PayloadAction<FetchVideosRequest>) => {
@@ -61,6 +65,7 @@ export const fetchesSlice = createSlice({
                 id: action.payload.channelId,
                 mode: action.payload.mode,
                 status: 'REQUESTED',
+                count: 0
             }
         },
         fetchedVideos: (state, action: PayloadAction<FetchVideosEvent>) => {
@@ -68,14 +73,15 @@ export const fetchesSlice = createSlice({
             const fetchState = state.videos[event.objectId];
             state.videos[event.objectId] = {
                 id: event.objectId,
-                status: event.status,
-                error: event.error,
+                status: 'FETCHING',
+                count: fetchState.count + event.items.length,
             }
         },
         requestedCommentsForVideoId: (state, action: PayloadAction<string>) => {
             state.comments[action.payload] = {
                 id: action.payload,
                 status: 'REQUESTED',
+                count: 0
             }
         },
         fetchedComments: (state, action: PayloadAction<FetchCommentsEvent>) => {
@@ -83,15 +89,15 @@ export const fetchesSlice = createSlice({
             const fetchState = state.comments[event.objectId];
             state.comments[event.objectId] = {
                 id: event.objectId,
-                status: event.status,
-                error: event.error,
+                status: 'FETCHING',
+                count: fetchState.count + event.items.length,
             }
         },
-        requestedRepliesForCommentId: (state, action: PayloadAction<string>) => {
+        requestedRepliesForId: (state, action: PayloadAction<string>) => {
             state.replies[action.payload] = {
                 id: action.payload,
                 status: 'REQUESTED',
-                data: []
+                count: 0
             }
         },
         fetchedReplies: (state, action: PayloadAction<FetchCommentsEvent>) => {
@@ -99,9 +105,17 @@ export const fetchesSlice = createSlice({
             const fetchState = state.replies[event.objectId];
             state.replies[event.objectId] = {
                 id: event.objectId,
+                status: 'FETCHING',
+                count: fetchState.count + event.items.length,
+            }
+        },
+        updateFetchStatus: (state, action: PayloadAction<UpdateFetchStatus>) => {
+            const event = action.payload;
+            const fetchState = state[event.fetchType][event.objectId];
+            state[event.fetchType][event.objectId] = {
+                id: event.objectId,
                 status: event.status,
-                data: (fetchState?.data ?? []).concat(event.items ?? []),
-                error: event.error,
+                count: fetchState.count
             }
         }
     }
@@ -114,8 +128,9 @@ export const {
     fetchedVideos,
     requestedCommentsForVideoId,
     fetchedComments,
-    requestedRepliesForCommentId,
+    requestedRepliesForId,
     fetchedReplies,
+    updateFetchStatus,
 } = fetchesSlice.actions;
 
 export default fetchesSlice.reducer;